Refactor: 예약 리포지토리 객체를 패키지로 관리하도록 수정
