+++ src/main/java/com/prgms/allen/dining/domain/reservation/repository/CustomReservationRepository.java
@@ -1,4 +1,4 @@
-package com.prgms.allen.dining.domain.reservation;
+package com.prgms.allen.dining.domain.reservation.repository;
 
 import com.prgms.allen.dining.domain.reservation.dto.CustomerReservationInfoParam;
 import com.prgms.allen.dining.domain.reservation.dto.CustomerReservationInfoProj;
